+++ daengle-domain/src/main/java/ddog/domain/user/User.java
@@ -1,12 +1,7 @@
 package ddog.domain.user;
 
-
 import ddog.domain.estimate.dto.response.UserInfo;
 import ddog.domain.pet.Pet;
-import ddog.domain.user.dto.request.JoinUserWithoutPet;
-import ddog.domain.user.dto.request.JoinUserWithPet;
-import ddog.domain.user.dto.response.PetInfo;
-import ddog.domain.user.dto.response.UserProfileInfo;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -32,42 +27,6 @@ public class User {
     private String address;
     private String email;
     private List<Pet> pets;
-
-    public static User createWithoutPet(Long accountId, JoinUserWithoutPet request) {
-        return User.builder()
-                .accountId(accountId)
-                .username(request.getUsername())
-                .phoneNumber(request.getPhoneNumber())
-                .nickname(request.getNickname())
-                .address(request.getAddress())
-                .email(request.getEmail())
-                .pets(new ArrayList<>())
-                .build();
-    }
-
-    public static User createWithPet(Long accountId, JoinUserWithPet request, Pet pet) {
-        List<Pet> pets = new ArrayList<>();
-        pets.add(pet);
-        return User.builder()
-                .accountId(accountId)
-                .username(request.getUsername())
-                .phoneNumber(request.getPhoneNumber())
-                .nickname(request.getNickname())
-                .address(request.getAddress())
-                .email(request.getEmail())
-                .pets(pets)
-                .build();
-    }
-
-    public UserProfileInfo toUserProfileInfo() {
-        return UserProfileInfo.builder()
-                .image(userImage)
-                .nickname(nickname)
-                .username(username)
-                .phoneNumber(phoneNumber)
-                .email(email)
-                .build();
-    }
 
     public User withImageAndNickname(String userImage, String nickname) {
         return User.builder()
@@ -95,29 +54,6 @@ public class User {
                 .address(address)
                 .email(email)
                 .pets(pets)
-                .build();
-    }
-
-    public PetInfo toPetInfo() {
-        List<PetInfo.Detail> petDetails = new ArrayList<>();
-        for (Pet pet : pets) {
-            petDetails.add(PetInfo.Detail.builder()
-                    .id(pet.getPetId())
-                    .image(pet.getPetImage())
-                    .name(pet.getPetName())
-                    .birth(pet.getPetBirth())
-                    .gender(pet.getPetGender())
-                    .breed(pet.getBreed())
-                    .isNeutered(pet.getIsNeutered())
-                    .weight(pet.getPetWeight())
-                    .groomingExperience(pet.getGroomingExperience())
-                    .isBite(pet.getIsBite())
-                    .dislikeParts(pet.getDislikeParts().split(","))
-                    .significant(pet.getPetSignificant())
-                    .build());
-        }
-        return PetInfo.builder()
-                .petDetails(petDetails)
                 .build();
     }
