--- daengle-user-api/src/main/java/ddog/user/presentation/dto/ValidNickname.java
@@ -1,12 +1,0 @@
-package ddog.user.presentation.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Builder
-@Getter
-public class ValidNickname {
-
-    private Boolean isAvailable;
-
-}
