+++ daengle-user-api/src/main/java/ddog/user/presentation/dto/response/ValidResponse.java
@@ -1,0 +1,15 @@
+package ddog.user.presentation.dto.response;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Builder
+@Getter
+public class ValidResponse {
+
+    @Builder
+    @Getter
+    public static class Nickname{
+        private Boolean isAvailable;
+    }
+}
